backend/requirements.txt
fastapi
uvicorn
python-jobspy
playwright
pydantic
python-multipart
pytest
httpx
langchain
langgraph
langchain-community
langchain-groq
tavily-python
pytest-asyncio
beautifulsoup4
python-dotenv
curl_cffi>=0.5.10
